+++ common/config.go
@@ -49,6 +49,7 @@ type RunnerConfig struct {
 
 	Shell          *string `toml:"shell" json:"shell"`
 	DisableVerbose *bool   `toml:"disable_verbose" json:"disable_verbose"`
+	OutputLimit    *int    `toml:"output_limit"`
 
 	SSH       *ssh.Config      `toml:"ssh" json:"ssh"`
 	Docker    *DockerConfig    `toml:"docker" json:"docker"`